Refactor mock files in deltaGenerator spec #2708 [Reduce code duplication for sonar qube]
